Arcigy_website/templates/scheduled_template.py
import os
import datetime
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

def run_scheduled_task():
    """
    Standard entry point for scheduled triggers.
    """
    try:
        now = datetime.datetime.now()
        print(f"Starting scheduled task at: {now}")
        
        # 1. Fetch data
        # 2. Perform enrichment/processing
        # 3. Log results
        
        return {"status": "success", "timestamp": str(now)}
        
    except Exception as e:
        print(f"Error in scheduled task: {e}")
        return {"status": "error", "message": str(e)}

if __name__ == "__main__":
    run_scheduled_task()
