requirements.txt
fastapi
uvicorn
retell-sdk
requests
python-dotenv
asyncio
pydantic
pydantic-settings
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
resend
jinja2
